Migrate PrimaryToolbar to TS
