+++ packages/components/src/BulkSelect/index.ts
@@ -1,2 +1,3 @@
 export { default } from './BulkSelect';
 export { default as BulkSelect } from './BulkSelect';
+export * from './BulkSelect';
