+++ packages/components/src/PrimaryToolbar/index.ts
@@ -3,3 +3,4 @@ export { default as PrimaryToolbar } from './PrimaryToolbar';
 export { default as Actions } from './Actions';
 export { default as SortBy } from './SortBy';
 export * from './SortBy';
+export * from './PrimaryToolbar';
